requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2.*
django-cors-headers==4.3.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
frozenlist==1.6.0
groq==0.26.0
gunicorn==21.2.*
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.4.4
openai==1.82.1
packaging==25.0
pillow>=10.1.0
propcache==0.3.1
psycopg2-binary==2.9.*
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
requests==2.32.3
setuptools==78.1.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1
whitenoise==6.5.*
yarl==1.20.0

# Testing dependencies
pytest==7.4.*
pytest-django==4.8.*
pytest-cov==4.0.*
factory-boy==3.3.*
faker==19.6.*
